+++ changelog.md
@@ -1,5 +1,6 @@
 # V0.7.2 -
 
+- 重写了 `Photos`，增加了手势缩放功能。
 - 新组件 `toptip`.
 - `Select` 增加了 `min`, `max`, `beforeClose` 三个配置参数，修复了在多选模式下取消全部选择后导致的Bug。
 - `ActionSheet` 增加了 `title` 参数，可以设置一个标题。